Added more robust testing at reasonable endpoints, and in the process:
Updated timeZoneOffsetMillis and its documentation to make coreJS and coreJVM return the same sign.
Updated dayOfWeek along with its documentation to return the day of the week 1-indexed instead of zero indexed, as per coreJVM.
Updated dayOfYear in JavaScriptCrossDate and JVMCrossDate to be zero-indexed as per the documentation.
Fixed a glitch in hour12 that caused it to be one hour off.
Fixed a small documentation error in def s.
Updated A and a to work with new dayOfWeek definition, since they are zero indexed.