Changed file name of coreNative's CrossDate implementation for clarity Began work on a fully scala-native implementation of CrossDate. Missing:
timeZone - there are other native classes (and I believe a locales class) that I think we can use here.

timeZoneOffsetMillis - there is a gmtime() method in posix that gives the UTC time of a given date, but it is based in seconds. A better approach would probably just be to determine the user's timezone using timeZone and figure out the UTC offset in milliseconds. The worse approach is commented out and does not work.